+++ Rebus.AzureStorage.Tests/Subscriptions/AzureSubscriptionStorageBasicSubscriptionOperations.cs
@@ -1,10 +1,10 @@
-using NUnit.Framework;
-using Rebus.Tests.Contracts.Subscriptions;
+//using NUnit.Framework;
+//using Rebus.Tests.Contracts.Subscriptions;
 
-namespace Rebus.AzureStorage.Tests.Subscriptions
-{
-    [TestFixture]
-    public class AzureSubscriptionStorageBasicSubscriptionOperations : BasicSubscriptionOperations<AzureStorageSubscriptionStorageFactory>
-    {
-    }
-}
+//namespace Rebus.AzureStorage.Tests.Subscriptions
+//{
+//    [TestFixture]
+//    public class AzureSubscriptionStorageBasicSubscriptionOperations : BasicSubscriptionOperations<AzureStorageSubscriptionStorageFactory>
+//    {
+//    }
+//}