+++ Rebus.AzureStorage/AzureStorage/AzureEx.cs
@@ -1,18 +1,18 @@
 ﻿using System.Net;
 using Microsoft.WindowsAzure.Storage;
-using Microsoft.WindowsAzure.Storage.Table;
+//using Microsoft.WindowsAzure.Storage.Table;
 using Rebus.Exceptions;
 
 namespace Rebus.AzureStorage
 {
     static class AzureEx
     {
-        public static void EnsureOk(this TableResult result)
-        {
-            if (result.HttpStatusCode == 200) return;
+        //public static void EnsureOk(this TableResult result)
+        //{
+        //    if (result.HttpStatusCode == 200) return;
 
-            throw new RebusApplicationException($"Table result status: {result.HttpStatusCode} - expected 200 OK");
-        }
+        //    throw new RebusApplicationException($"Table result status: {result.HttpStatusCode} - expected 200 OK");
+        //}
 
         public static bool IsStatus(this StorageException exception, HttpStatusCode statusCode)
         {