+++ Rebus.AzureStorage/AzureStorage/Subscription.cs
@@ -1,15 +1,15 @@
-﻿using Microsoft.WindowsAzure.Storage.Table;
+﻿//using Microsoft.WindowsAzure.Storage.Table;
 
-namespace Rebus.AzureStorage
-{
-    class AzureStorageSubscription : TableEntity
-    {
-        public AzureStorageSubscription(string topic, string subscriberAddress)
-        {
-            PartitionKey = topic;
-            RowKey = subscriberAddress;
-        }
+//namespace Rebus.AzureStorage
+//{
+//    class AzureStorageSubscription : TableEntity
+//    {
+//        public AzureStorageSubscription(string topic, string subscriberAddress)
+//        {
+//            PartitionKey = topic;
+//            RowKey = subscriberAddress;
+//        }
 
-        public AzureStorageSubscription() { }
-    }
-}
+//        public AzureStorageSubscription() { }
+//    }
+//}